split UserGuide into smaller components and added docstrings to /UserGuide/ components
